#33961 ブランチの作成。ただし通常のブランチとしてではなく、DTXMania(Release099)からの派生である。

+++ SlimDXc_Jun2010(VC++2008)/source/xaudio2/PerformanceData.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 /*
-* Copyright (c) 2007-2012 SlimDX Group
+* Copyright (c) 2007-2010 SlimDX Group
 * 
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -31,8 +31,8 @@ namespace XAudio2
 {
 	PerformanceData::PerformanceData( const XAUDIO2_PERFORMANCE_DATA &data )
 	{
-		AudioCyclesSinceLastQuery = data.AudioCyclesSinceLastQuery;
-		TotalCyclesSinceLastQuery = data.TotalCyclesSinceLastQuery;
+		AudioCyclesSinceLastQuery = static_cast<long>( data.AudioCyclesSinceLastQuery );
+		TotalCyclesSinceLastQuery = static_cast<long>( data.TotalCyclesSinceLastQuery );
 		MinimumCyclesPerQuantum = data.MinimumCyclesPerQuantum;
 		MaximumCyclesPerQuantum = data.MaximumCyclesPerQuantum;
 		MemoryUsageInBytes = data.MemoryUsageInBytes;
